pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multitabfm"
version = "0.1.0"
description = "Multi-table feature modeling framework with AutoGluon integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AWS Shanghai AI Lab"}
]
maintainers = [
    {name = "AWS Shanghai AI Lab"}
]
keywords = ["machine learning", "feature engineering", "autogluon", "multi-table", "dfs", "tabular data"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"

# Core dependencies
dependencies = [
    "pandas",
    "numpy",
    "PyYAML",
    "scikit-learn",
    "autogluon.tabular>=1.0.0",
    "tabpfn==2.1.3",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

# Testing dependencies
testing = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]


[project.urls]
Homepage = "https://github.com/your-username/multitabfm"
Documentation = "https://multitabfm.readthedocs.io/"
Repository = "https://github.com/your-username/multitabfm.git"
"Bug Tracker" = "https://github.com/your-username/multitabfm/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["multitabfm*"]

[tool.setuptools.package-data]
multitabfm = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38']


[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=multitabfm --cov-report=term-missing --cov-report=html"